code_cleanup_0
Cleaned up code to make it more pythonic and edited variable names.

Cleaning is still in progress
